Move parse_value and SIMULATION_TIMEOUT to shared utils package (#568)
The simulation layer (power_calculator, power_metrics, ngspice_runner)
was importing parse_value from GUI.format_utils and SIMULATION_TIMEOUT
from GUI.styles, creating a backwards dependency on the GUI layer.

- Create app/utils/ package with format_utils.py and constants.py
- Move parse_value, format_value, validate_component_value to utils.format_utils
- Move SIMULATION_TIMEOUT to utils.constants
- Update simulation imports to use utils instead of GUI
- Keep GUI modules as re-export shims for backward compatibility
