+++ app/utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+Shared utilities for the Spice-GUI application.
+
+This package contains pure-Python helpers that may be used by any layer
+(models, controllers, simulation, GUI) without introducing cross-layer
+dependencies.
+"""
